Refactor Atlas Game structure and remove unused classes
