+++ main/java/AtlasExceptions.java
@@ -1,4 +1,4 @@
-package main;
+package main.java;
 public class AtlasExceptions {
     
     public static class AtlasGameException extends Exception {
