+++ main/java/AtlasGameHistory.java
@@ -1,4 +1,4 @@
-package main;
+package main.java;
 import java.util.ArrayList;
 import java.util.List;
