+++ main/java/AtlasPlaceValidator.java
@@ -1,4 +1,4 @@
-package main;
+package main.java;
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
