+++ test/java/TestRunner.java
@@ -1,4 +1,4 @@
-package test;
+package test.java;
 public class TestRunner {
     public static class AssertBooleansException extends Exception {
         public AssertBooleansException(String message,Object actual, Object expected) {
